+++ library/Edge.h
@@ -4,15 +4,15 @@
 typedef struct edge Edge;
 typedef Edge** EdgesArray;
 
-Edge* init_edge(int id_a, int id_b, double weight);
+Edge* init_edge(unsigned  short id_a, unsigned short id_b, float weight);
 
-double get_weight_edge(Edge* e);
+float get_weight_edge(Edge* e);
 
 void print_edge(Edge* e);
 
-int get_id_a_edge(Edge* e);
+unsigned short get_id_a_edge(Edge* e);
 
-int get_id_b_edge(Edge * e);
+unsigned short get_id_b_edge(Edge * e);
 
 int cmp_edge(const void *a, const void *b);
