+++ library/Graph.h
@@ -10,7 +10,7 @@ void add_vertex_to_graph(Graph * g, Vertex * v);
 
 void add_to_adjacency_list_of(Vertex * v1, Vertex * v2);
 
-int get_vertex_id(Vertex* v);
+unsigned short get_vertex_id(Vertex* v);
 
 void print_graph_elements(Graph* g);
