+++ library/PlanesPoint.c
@@ -5,13 +5,13 @@
 #include "PlanesPoint.h"
 
 struct planesPoint{
-    double tuple[2];
-    int id;
+    float tuple[2];
+    unsigned short id;
 };
 
 
 
-PlanesPoint * init_planes_point(int id, double x, double y){
+PlanesPoint * init_planes_point(unsigned short id, float x, float y){
     PlanesPoint* p = (PlanesPoint*)malloc(sizeof(PlanesPoint));
     if(p == NULL){
         fprintf(stderr, "ERROR, ALLOCATION OF P FAILED!\n");
@@ -24,11 +24,11 @@ PlanesPoint * init_planes_point(int id, double x, double y){
     return p;
 }
 
-double get_x_planes_point(PlanesPoint* p){
+float get_x_planes_point(PlanesPoint* p){
     return p->tuple[0];
 }
 
-double get_y_planes_point(PlanesPoint* p){
+float get_y_planes_point(PlanesPoint* p){
     return p->tuple[1];
 }
 
@@ -39,11 +39,11 @@ void free_planes_point(PlanesPoint* p){
 }
 
 void print_planes_point(PlanesPoint* p){
-    printf("{\n  id = %d\n  x = %lf\n  y = %lf\n}\n", p->id, p->tuple[0], p->tuple[1]);
+    printf("{\n  id = %d\n  x = %f\n  y = %f\n}\n", p->id, p->tuple[0], p->tuple[1]);
 }
 
-double distance_between_planes_point(PlanesPoint* p1, PlanesPoint* p2){
-    double distance  = (p1->tuple[0] - p2->tuple[0])*(p1->tuple[0] - p2->tuple[0]) + (p1->tuple[1] - p2->tuple[1])*(p1->tuple[1] - p2->tuple[1]);
+float distance_between_planes_point(PlanesPoint* p1, PlanesPoint* p2){
+    float distance  = (p1->tuple[0] - p2->tuple[0])*(p1->tuple[0] - p2->tuple[0]) + (p1->tuple[1] - p2->tuple[1])*(p1->tuple[1] - p2->tuple[1]);
 
     return sqrt(distance);
 }
@@ -56,6 +56,6 @@ void copy_to_planes_point(PlanesPoint * p1, PlanesPoint* p2){
     return;
 }
 
-int get_id_planes_point(PlanesPoint* p){
+unsigned short get_id_planes_point(PlanesPoint* p){
     return p->id;
 }
