+++ library/UnionFind.h
@@ -4,16 +4,16 @@
 typedef struct uf UF;
 
 // incia o conjunto/árvore/grafo que utiliza do algoritimo de Union-Find, os elementos são representados por numeros que começam do 1 e vão até N (inclusivo)
-UF* initUnionFind(int n);
+UF* initUnionFind(unsigned short n);
 
 //une dois elementos que são representados por inteiros na estrutura
-void Union(UF* uf,  int a,int b);
+void Union(UF* uf,  unsigned short a, unsigned short b);
 
 //verifica se dois elementos estao conectados)
-int Conected(UF* uf, int a, int b);
+unsigned short Conected(UF* uf, unsigned short a, unsigned short b);
 
 //acha o pai do elemento especificado
-int Find(UF* uf, int a);
+unsigned short Find(UF* uf, unsigned short a);
 
 //libera toda a estrutura do union find
 void freeUnionFind(UF* uf);
